Change nne, TR, TJ, W from double to float (including radfield bins)

+++ grid_init.h
@@ -98,22 +98,22 @@ inline float get_kappagrey(int modelgridindex)
   return modelgrid[modelgridindex].kappagrey;
 }
 
-inline double get_Te(int modelgridindex)
+inline float get_Te(int modelgridindex)
 {
   return modelgrid[modelgridindex].Te;
 }
 
-inline double get_TR(int modelgridindex)
+inline float get_TR(int modelgridindex)
 {
   return modelgrid[modelgridindex].TR;
 }
 
-inline double get_TJ(int modelgridindex)
+inline float get_TJ(int modelgridindex)
 {
   return modelgrid[modelgridindex].TJ;
 }
 
-inline double get_W(int modelgridindex)
+inline float get_W(int modelgridindex)
 {
   return modelgrid[modelgridindex].W;
 }
